LayoutTests/fast/svg/foreign-object-offset-position-expected.txt

PASS event.offsetX is 10
PASS event.offsetY is 10
PASS successfullyParsed is true

TEST COMPLETE
